Move some routes from badge/routes.py to groups.py and do the same with corresponding server tests
