+++ timApp/tests/server/test_groups.py
@@ -651,6 +651,239 @@ class GroupNotificationsTest(TimRouteTest):
         )
 
 
+class SubgroupsTest(TimRouteTest):
+    def test_subgroups(self):
+        # initialization
+        self.login_test1()
+        group1_name = "es_25"
+        subgroup1_name = "es_25-cats"
+        subgroup2_name = "es_25-dogs"
+        group2_name = "es_26"
+        (group1, doc1) = do_create_group_impl(f"{group1_name}", group1_name)
+        (subgroup1, subdoc1) = do_create_group_impl(f"{subgroup1_name}", subgroup1_name)
+        (subgroup2, subdoc2) = do_create_group_impl(f"{subgroup2_name}", subgroup2_name)
+        (group2, doc2) = do_create_group_impl(f"{group2_name}", group2_name)
+        db.session.commit()
+        self.test_user_1.grant_access(group1.admin_doc, AccessType.teacher)
+        self.test_user_1.grant_access(subgroup1.admin_doc, AccessType.teacher)
+        self.test_user_1.grant_access(subgroup2.admin_doc, AccessType.teacher)
+        self.test_user_1.grant_access(group2.admin_doc, AccessType.teacher)
+        self.commit_db()
+
+        # fetch subgroups when there are not any
+        result_sg_empty = self.get(f"/groups/subgroups/{group2_name}")
+        self.assertEqual([], result_sg_empty)
+
+        # fetch subgroups when there are 2 of them
+        result_sg_nonempty = self.get(f"/groups/subgroups/{group1_name}")
+        self.assertEqual(
+            [
+                {"id": 10, "name": subgroup1_name},
+                {"id": 11, "name": subgroup2_name},
+            ],
+            result_sg_nonempty,
+        )
+
+        self.login_test2()
+
+        # fetch subgroups when user doesn't have teacher access to the context group
+        self.get(
+            f"/groups/subgroups/{group1_name}",
+            expect_content=f'Sorry, you don\'t have permission to use this resource. If you are a teacher of "{group1_name}", please contact TIM admin.',
+            expect_status=403,
+        )
+
+        self.login_test1()
+
+        # fetch subgroups with erroneous data
+        self.get(
+            f"/groups/subgroups/nonexistent_group",
+            expect_status=404,
+            expect_content='User group "nonexistent_group" not found',
+        )
+
+
+class UsersSubgroupsTest(TimRouteTest):
+    def test_users_subgroups(self):
+        # initialization
+        self.login_test1()
+        group1_name = "es_25"
+        subgroup1_name = "es_25-cats"
+        subgroup2_name = "es_25-dogs"
+        (group1, doc1) = do_create_group_impl(f"{group1_name}", group1_name)
+        (subgroup1, subdoc1) = do_create_group_impl(f"{subgroup1_name}", subgroup1_name)
+        (subgroup2, subdoc2) = do_create_group_impl(f"{subgroup2_name}", subgroup2_name)
+        db.session.commit()
+        self.test_user_1.grant_access(group1.admin_doc, AccessType.teacher)
+        self.test_user_1.grant_access(subgroup1.admin_doc, AccessType.teacher)
+        self.test_user_1.grant_access(subgroup2.admin_doc, AccessType.teacher)
+        self.commit_db()
+        self.post(
+            f"/groups/addmember/{group1_name}",
+            data={"names": ["testuser2", "testuser3"]},
+        )
+        self.post(
+            f"/groups/addmember/{subgroup1_name}",
+            data={"names": ["testuser2"]},
+        )
+
+        # fetch users subgroups when there are not any
+        result_usg_empty = self.get(
+            f"/groups/users_subgroups/{self.test_user_3.id}/{group1_name}"
+        )
+        self.assertEqual([], result_usg_empty)
+
+        # fetch users subgroups when user has teacher access to the context group
+        # and is not the user with given user id
+        self.get(
+            f"/groups/users_subgroups/{self.test_user_2.id}/{group1_name}",
+            expect_content=[{"id": 10, "name": subgroup1_name}],
+            expect_status=200,
+        )
+
+        self.login_test2()
+
+        # fetch users subgroups when user doesn't have teacher access to the context group
+        # and is the user with given user id
+        self.get(
+            f"/groups/users_subgroups/{self.test_user_2.id}/{group1_name}",
+            expect_content=[{"id": 10, "name": subgroup1_name}],
+            expect_status=200,
+        )
+
+        self.login_test3()
+
+        # fetch users subgroups when user doesn't have teacher access to the context group
+        # and is not the user with given user id
+        self.get(
+            f"/groups/users_subgroups/{self.test_user_2.id}/{group1_name}",
+            expect_content=f'Sorry, you don\'t have permission to use this resource. If you are a teacher of "{group1_name}", please contact TIM admin.',
+            expect_status=403,
+        )
+
+        self.login_test1()
+
+        # fetch users subgroups with different erroneous data
+        self.get(
+            f"/groups/users_subgroups/100/{group1_name}",
+            expect_status=404,
+            expect_content='User with id "100" not found',
+        )
+        self.get(
+            f"/groups/users_subgroups/{self.test_user_2.id}/nonexistent_group",
+            expect_status=404,
+            expect_content='User group "nonexistent_group" not found',
+        )
+
+
+class UserAndPersonalGroupTest(TimRouteTest):
+    def test_user_and_personal_group(self):
+        # initialization
+        self.login_test2()
+
+        # fetch user and his/her personal usergroup
+        result_uapg = self.get(
+            f"/groups/user_and_personal_group/{self.test_user_2.name}"
+        )
+        self.assertEqual(
+            [
+                {
+                    "id": self.test_user_2.id,
+                    "name": self.test_user_2.name,
+                    "real_name": self.test_user_2.real_name,
+                    "email": self.test_user_2.email,
+                },
+                {
+                    "id": self.test_user_2.get_personal_group().id,
+                    "name": self.test_user_2.get_personal_group().name,
+                },
+            ],
+            result_uapg,
+        )
+
+        self.login_test1()
+
+        # fetch user and his/her personal usergroup with erroneous data
+        self.get(
+            f"/groups/user_and_personal_group/nonexistent_user",
+            expect_status=404,
+            expect_content='User "nonexistent_user" not found',
+        )
+
+
+class UsergroupsMembersTest(TimRouteTest):
+    def test_usergroups_members(self):
+        # initialization
+        self.login_test1()
+        subgroup1_name = "es_25-cats"
+        subgroup2_name = "es_25-dogs"
+        (subgroup1, subdoc1) = do_create_group_impl(f"{subgroup1_name}", subgroup1_name)
+        (subgroup2, subdoc2) = do_create_group_impl(f"{subgroup2_name}", subgroup2_name)
+        db.session.commit()
+        self.test_user_1.grant_access(subgroup1.admin_doc, AccessType.teacher)
+        self.test_user_1.grant_access(subgroup2.admin_doc, AccessType.teacher)
+        self.commit_db()
+        self.post(
+            f"/groups/addmember/{subgroup1_name}",
+            data={"names": ["testuser2"]},
+        )
+
+        # fetch usergroup's members when there are not any
+        result_ugm_empty = self.get(f"/groups/usergroups_members/{subgroup2_name}")
+        self.assertEqual(
+            [],
+            result_ugm_empty,
+        )
+
+        # fetch usergroup's members when there are 2
+        result_ugm_nonempty = self.get(f"/groups/usergroups_members/{subgroup1_name}")
+        self.assertEqual(
+            [
+                {
+                    "id": self.test_user_2.id,
+                    "name": self.test_user_2.name,
+                    "real_name": self.test_user_2.real_name,
+                    "email": self.test_user_2.email,
+                }
+            ],
+            result_ugm_nonempty,
+        )
+
+        self.login_test2()
+
+        # fetch usergroup's members when user doesn't have view access to the usergroup and belongs to usergroup
+        self.get(
+            f"/groups/usergroups_members/{subgroup1_name}",
+            expect_content=[
+                {
+                    "id": self.test_user_2.id,
+                    "name": self.test_user_2.name,
+                    "real_name": self.test_user_2.real_name,
+                    "email": self.test_user_2.email,
+                }
+            ],
+            expect_status=200,
+        )
+
+        self.login_test3()
+
+        # fetch usergroup's members when user doesn't have view access to the usergroup and doesn't belong to usergroup
+        self.get(
+            f"/groups/usergroups_members/{subgroup1_name}",
+            expect_content=f"Sorry, you don't have permission to use this resource.",
+            expect_status=403,
+        )
+
+        self.login_test1()
+
+        # fetch usergroup's members with erroneous data
+        self.get(
+            f"/groups/usergroups_members/nonexistent_group",
+            expect_status=404,
+            expect_content='User group "nonexistent_group" not found',
+        )
+
+
 class GroupNameChangerTest(TimRouteTest):
     def test_group_name_changer(self):
         # initialization
